coinalyze/requirements.txt
alembic==1.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
gate-api==4.57.0
gate-ws==0.3.1
gevent==23.9.1
greenlet==3.0.1
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
pika==1.3.2
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
websocket-client==1.6.4
websockets==12.0
zope.event==5.0
zope.interface==6.1